More refactoring.
I wanted to showcase `cargo doc`, but there were complaints about having both a lib and a bin in the same package, so I moved the lib to a separate crate.

I also moved the coordinates logic to its own module, preparing for the parser.

There are some tests too.