Implemented the core of the spreadsheet.
We have a `HashMap<Coord, Formula>` that holds the formulas in each coord.

We can then ask for the value of some coordinates. The thing will interpret the formula held in the hash map and return a `FormulaAtom`, which can only be string, number, of empty.

`Sheet::value` returns `Result<Box<FormulaAtom>, FormulaErr>`. Why box? Well, I started with plain `FormulaAtom`, but the strings held inside `Formula::String` were problematic, of course. That could've solved with `.clone()`, I think. Anyway, I moved on to `&FormulaAtom`, but, of course, when doing operations (sums, etc.) the thing needs to make new `FormulaAtom`, so in order to return that new ones they need to be in the heap. So I'm pushing quite a bit the heap for this instead of copying/cloning things around. Not sure which is best. You have to think about this kind of things in Rust.

Note a beautiful piece of code reuse with `Sheet::numeric_op`. You can use a generic function that uses a method as a generic parameter with zero runtime overhead.

So, what's left: the parser and formatter of formulae; a channel that takes editing events from the UI to the sheet; a channel that informs of changes in a selection of the sheet to the UI, to re-render; moving around the sheet from the UI; I think that's pretty much it.