The parser is ready, kind of.
I've used a PEG parser generator. It uses a syntax extension for compiling on the fly with the rest of the program, so no need for a external command. That's pretty awesome. Regexps, SQL, etc. can work the same way. The library is pretty easy, too.

It was _really_ late when I did `numeric_col_to_natural`, so, it's a bit broken.